Otimiza consulta por pessoa nas condições: (#252)
- Consulta genérica por nome LIKE
- Consulta com condicional de filtragem para pessoas já "fechadas" no sistema (dataFimPessoa <> null);
- Consulta com condicional para incluir substitutos;

+++ siga-cp/src/main/java/br/gov/jfrj/siga/cp/bl/CpBL.java
@@ -1290,7 +1290,7 @@ public class CpBL {
 		dpPessoa.setId(id);
 
 		dpPessoa.setBuscarFechadas(Boolean.FALSE);
-		Integer tamanho = CpDao.getInstance().consultarQuantidade(dpPessoa);
+		long tamanho = CpDao.getInstance().consultarQuantidade(dpPessoa);
 
 		if (tamanho > 0) {
 			throw new AplicacaoException("Usuário já cadastrado com estes dados: Órgão, Cargo, Função, Unidade e CPF");
